Change: refactor for easier testing
Pulling out the generation of the url and options makes the code much
easier to read and test, and rather than messing about with trying to
set environment veriables, we can just use a test double by getting the
getenv call to return what we want.  We can use this when we test the
sad path where the cookie is bad or not there.